refactor(config): extract useConfigScreenState and configHelpers from ConfigScreen
